move question schema doc to meta db
