+++ corehq/couchapps/__init__.py
@@ -1,3 +1,5 @@
 from corehq.preindex import CouchAppsPreindexPlugin
 
-CouchAppsPreindexPlugin.register('couchapps', __file__)
+CouchAppsPreindexPlugin.register('couchapps', __file__, {
+    'form_question_schema': 'meta'
+})
