+++ settings.py
@@ -1031,6 +1031,7 @@ COUCHDB_APPS = [
     ('mc', 'fluff-mc'),
     ('m4change', 'm4change'),
     ('wvindia2', 'wvindia2'),
+    ('export', 'meta'),
     'tdhtesting'
 ]
